requirements.txt
pdfminer.six
pdfplumber
docx2txt
python-docx
numpy
pandas
scikit-learn
nltk
spacy
sentence-transformers
tqdm
joblib
flask 
flask-cors 
python-dotenv 
qdrant-client 
sentence-transformers 
pandas
